Restructure package for clarity
