+++ mcp_project/__init__.py
@@ -1,6 +1,10 @@
-"""MCP Project example package."""
+"""MCP Project example package.
+
+This file exposes the most useful functions so importing ``mcp_project`` feels
+similar to opening a toolbox and seeing all the common tools ready to use.
+"""
 
 __all__ = ["main", "ProjectTree"]
 
 from .cli import main
-from .tree import ProjectTree
+from .core import ProjectTree
