+++ mcp_project/core/__init__.py
@@ -1,0 +1,5 @@
+"""Core functionality for MCP Project."""
+
+from .tree import ProjectTree
+
+__all__ = ["ProjectTree"]
