+++ mcp_project/tools/__init__.py
@@ -1,0 +1,6 @@
+"""Additional tools that go beyond the basics."""
+
+# Expose TreeGenerator for convenience
+from .tree_generator import ProjectTreeGenerator
+
+__all__ = ["ProjectTreeGenerator"]
