+++ pyproject.toml
@@ -8,3 +8,9 @@ version = "0.1.0"
 description = "Example project skeleton"
 readme = "README.md"
 requires-python = ">=3.11"
+
+[tool.setuptools.packages.find]
+where = ["."]
+
+[project.scripts]
+mcp-project = "mcp_project.cli:main"
